Enforce auth for server side actions (#69)
* Enforce most of the server actions auth

Also adds a toast that shows the result of a request

* Add auth for more pages

* Update locale for some toasts and pages

* Add remaining auth enforcement

* Fix formating

* Use general i18n

* More formating fix

+++ src/actions/banners.ts
@@ -4,8 +4,18 @@ import MediaService from '@/services/mediaService';
 import DivisionGroupService from '@/services/divisionGroupService';
 import { redirect } from 'next/navigation';
 import { MediaType } from '@/services/fileService';
+import SessionService from '@/services/sessionService';
 
 export async function addBanner(divisionGroupId: number, form: FormData) {
+  if (
+    !(
+      (await SessionService.isAdmin()) ||
+      (await SessionService.canEditGroupByInternalId(divisionGroupId))
+    )
+  ) {
+    throw new Error('Unauthorized');
+  }
+
   const file: File | null = form.get('file') as unknown as File;
 
   if (!file) {
@@ -20,6 +30,20 @@ export async function addBanner(divisionGroupId: number, form: FormData) {
 }
 
 export async function deleteBanner(bannerId: number) {
+  const divisionGroupId = await DivisionGroupService.getBannerOwner(bannerId);
+  if (divisionGroupId === undefined) {
+    throw new Error('Banner owner not found');
+  }
+
+  if (
+    !(
+      (await SessionService.isAdmin()) ||
+      (await SessionService.canEditGroupByInternalId(divisionGroupId))
+    )
+  ) {
+    throw new Error('Unauthorized');
+  }
+
   await DivisionGroupService.deleteBanner(bannerId);
   redirect('./banners');
 }
